+++ hyperdbg/hyperkd/header/debugger/kernel-level/Kd.h
@@ -233,3 +233,6 @@ KdCheckTheHaltedCore(PROCESSOR_DEBUGGING_STATE * DbgState);
 BOOLEAN
 KdQueryDebuggerQueryThreadOrProcessTracingDetailsByCoreId(UINT32                          CoreId,
                                                           DEBUGGER_THREAD_PROCESS_TRACING TracingType);
+
+BOOLEAN
+KdQueryIgnoreHandlingMov2DebugRegs(UINT32 CoreId);